libc: strtof, strtold, atoi, puts and strlen are declared with the standard's types
A program may declare a library function itself, and C99 6.7p4 requires
that declaration to agree with the header's. Several did not:

- `strtof` returned `double` and was bound to `strtod`, where C99 7.20.1.3
  declares `float strtof(const char *, char **)`. libSystem and glibc
  export `strtof` and the binding now names it; the glibc symbol-version
  manifest gains its entry, at the baseline version its neighbours
  `strtod` and `strtold` carry on each architecture. msvcrt.dll exports no
  `strtof`, so on Windows the header defines it over `strtod`.
- `strtold` and `atoi` took `char *`, and `puts` did too, where C99 takes
  `const char *`.
- `strlen` was declared `int strlen(char *)`, which truncates a length past
  2^31. It returns `size_t`; the snapshots of the fixtures that call it
  drop the 32-bit extension of the result and compare it unsigned.

glibc declares `mempcpy` under `_GNU_SOURCE`; libSystem and msvcrt do not
declare it at all. <string.h> defined it on every target, so a program
that supplies its own where the platform lacks one -- nasm does, from its
configuration -- defined the function twice. Every unit that includes the
header also carried the unused body, which shifts the entry numbers in the
SSA snapshots.

block_scope_function_declaration.c declares `strlen` with `size_t`, which
is the standard's type on every target; `unsigned long` is narrower than
`size_t` on Windows.

A test redeclares `popen`, `fdopen`, `strtof`, `strtold`, `strlen`, `puts`,
`atoi` and the wide-string functions with the standard's prototypes after
the headers, and defines `mempcpy`, on every target.

+++ tests/fixtures/c/block_scope_function_declaration.c
@@ -46,7 +46,7 @@ int main(void) {
     // binding -- the call still resolves through libc, not as an
     // undefined user reference.
     {
-        extern unsigned long strlen(const char *s);
+        extern size_t strlen(const char *s);
         if (strlen("abcd") != 4) {
             return 4;
         }
